fix bug

+++ utilCartoon.py
@@ -103,7 +103,7 @@ for idx, image_path in enumerate(framePaths) :
     image = load_image(image_path)
     kMaxTryCount = 3
     for tryIdx in range(kMaxTryCount):
-        api_logger.info(f"卡通化 {image_path}, 第{tryIdx + 1}生成")
+        api_logger.info(f"卡通化 {image_path}, 第{tryIdx + 1}次生成")
         image = pipeline("Cartoonize the following image", 
                         image=image,
                         num_inference_steps=20,
@@ -117,11 +117,11 @@ for idx, image_path in enumerate(framePaths) :
         fileSize = os.path.getsize(cartoonImagePath) 
         kRetryFileSizeK = 10 * 1024
         kMinFileSizeK = 2 * 1024
-        if os.path.getsize(cartoonImagePath) < kRetryFileSizeK :
-            api_logger.error("文件小于10K，生成错误，重试")
+        if fileSize < kRetryFileSizeK :
+            api_logger.error(f"文件 {fileSize} 小于10K，生成错误，重试")
             if tryIdx == kMaxTryCount - 1:
                 api_logger.info(f"已经重试{kMaxTryCount}次, 不再重试。")
-                if fileSize < kMinFileSizeK:
+                if os.path.exists(cartoonImagePath) and  fileSize < kMinFileSizeK:
                     api_logger.info(f"文件小于 {kMinFileSizeK} Byte, 将会删除 {cartoonImagePath}")
                     os.remove(cartoonImagePath)
             continue
